Publish completed order to queue
After an order is completed, publish the order details.
